Re-implement DuplicateObjectException
