+++ src/Exception/BadRequestException.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace App\Exception;
 
-use App\ErrorResponse\FooErrorInterface;
+use App\ErrorResponse\BadRequestErrorInterface;
 
 class BadRequestException extends FooException
 {
-    public function __construct(FooErrorInterface $error)
+    public function __construct(BadRequestErrorInterface $error)
     {
         parent::__construct($error, 400);
     }
