+++ src/Exception/DuplicateObjectException.php
@@ -5,32 +5,11 @@ declare(strict_types=1);
 namespace App\Exception;
 
 use App\ErrorResponse\DuplicateObjectErrorInterface;
-use App\RequestField\FieldInterface;
 
-/**
- * @phpstan-import-type SerializedDuplicateObjectError from DuplicateObjectErrorInterface
- */
-class DuplicateObjectException extends ErrorException implements DuplicateObjectErrorInterface
+class DuplicateObjectException extends FooException
 {
-    public function __construct(
-        private readonly FieldInterface $field,
-    ) {
-        parent::__construct(DuplicateObjectErrorInterface::ERROR_CLASS, null, '', 400);
-    }
-
-    public function getField(): FieldInterface
+    public function __construct(DuplicateObjectErrorInterface $error)
     {
-        return $this->field;
-    }
-
-    /**
-     * @return SerializedDuplicateObjectError
-     */
-    public function serialize(): array
-    {
-        return [
-            'class' => DuplicateObjectErrorInterface::ERROR_CLASS,
-            'field' => $this->field->jsonSerialize(),
-        ];
+        parent::__construct($error, 400);
     }
 }
